🛠 Fix: frame can take float value

+++ src/muml/rectangle.nim
@@ -1,12 +1,12 @@
-import json, types, utils, Palette/color
+import json, types, utils
 
 proc getRectangle* (muml: mumlNode): mumlObject =
   result = mumlObject(kind: mumlKindRectangle)
   for key, val in muml.pairs:
     case key:
     of "frame":
-      result.frame.start = val["start"].getFloat
-      result.frame.`end` = val["end"].getFloat
+      result.frame.start = val["start"].getInt
+      result.frame.`end` = val["end"].getInt
     of "layer": result.layer = val.getInt
     of "position": result.rectangle.position = val.getPosition
     of "width": result.rectangle.width = val.getNumberValue
